# docs/api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Atriom",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "https://mall-management.test/api"
        }
    ],
    "paths": {
...
        "/v1/me/requests/{id}/comments": {
            "post": {
                "operationId": "v1.me.requests.comment",
                "tags": [
                    "CommentTenantRequest"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CommentTenantRequestRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/TenantRequestCommentResource"
                                        },
                                        "message": {
                                            "type": "string",
                                            "const": "Comment added."
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CommentTenantRequestRequest": {
                "type": "object",
                "properties": {
                    "body": {
                        "type": "string",
                        "maxLength": 5000
                    }
                },
                "required": [
                    "body"
                ],
                "title": "CommentTenantRequestRequest"
            },
...
            "TenantRequestCommentResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "body": {
                        "type": "string"
                    },
                    "authorKind": {
                        "type": "string",
                        "enum": [
                            "tenant",
                            "staff"
                        ]
                    },
                    "authorName": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "string",
                                "enum": [
                                    "Property team"
                                ]
                            }
                        ]
                    },
                    "createdAt": {
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "body",
                    "authorKind",
                    "authorName",
                    "createdAt"
                ],
                "title": "TenantRequestCommentResource"
            },
...
        },
        "responses": {
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            },
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
...
        }
    }
}